Fixing partialSums, adding print_array, printing the input
